# account_manager/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: vasco.m.campos@inesctec.pt
    name: Vasco Campos
  description: 'User Account Manager Service OpenAPI definition. This service has
    the following functions: register, login, change user credentials and update it''s
    profile settings. Find out more: [User Account Manager Service documentation](https://gitlab.inesctec.pt/cpes/european-projects/interconnect/hems/hems-documentation/-/blob/master/Microservices/User-Account-Manager-Service.adoc)'
  title: User Account Manager Service
  version: 1.2.0
servers:
- url: https://interconnect-dev.inesctec.pt/api/account
security:
- {}
- Bearer: []
...
paths:
  /app/info:
    description: Info messages to be displayed in the app
    get:
      operationId: app_info_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: system_language
        required: true
        schema:
          enum:
          - pt-PT
          - en-GB
          example: pt-PT
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/InfoMessages'
                type: array
          description: List of info messages to show the user
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Bad format:
                  value:
                    error: Invalid specification format
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "404":
          content:
            application/json:
              examples:
                User Id not found:
                  value:
                    error: user with id "userId" not found
              schema:
                $ref: '#/components/schemas/Error'
          description: User id not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Info messages to be displayed in the app
      tags:
      - App - User Communication
      x-openapi-router-controller: account_manager.controllers.app__user_communication_controller
...
components:
  headers:
    CorrelationId:
      explode: false
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationId'
      style: simple
...
  schemas:
    CorrelationId:
      format: uuid
      type: string
...
    Authorization:
      format: byte
      type: string
...
    InfoMessages:
      example:
        icon: cloud-off
        description: The service is down for maintenance. Therefore, the data used
          will be profiled according to historical data.
        title: Smart meter service in maintenance
      properties:
        title:
          example: Smart meter service in maintenance
          title: title
          type: string
        description:
          example: The service is down for maintenance. Therefore, the data used will
            be profiled according to historical data.
          title: description
          type: string
        icon:
          example: cloud-off
          title: icon
          type: string
      title: InfoMessages
      type: object
    Error:
      description: Error model
      properties:
        error:
          title: error
          type: string
      required:
      - error
      title: Error
      type: object
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: account_manager.controllers.security_controller_.info_from_Bearer